% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{rpxCache}
\alias{rpxCache}
\alias{pxCachedProjects}
\title{Package cache}
\usage{
rpxCache()

pxCachedProjects(cache = rpxCache())
}
\arguments{
\item{cache}{Object of class `BiocFileCache`.}
}
\value{
`character(1)` with the path to the cache directory for
    `rpxCache()` and a `tibble` for `pxCachedProjects()`.
}
\description{
Function to access the cache directory. `rxpCache()` returns the
central `rpx` cache. `pxCachedrojects()` prints the names of the
cached projects and invisibly returns the cache table.
}
\details{
The cache is an object of class `BiocFileCache`, and created with
`BiocFileCache::BiocFileCache()`. It can be either the
package-wide cache as defined by `rpxCache()` or an instaned
provided by the user.
}
\examples{

rpxCache()

pxCachedProjects()
}
\author{
Laurent Gatto
}
