% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_types.R
\name{file_types}
\alias{file_types}
\alias{px_file_types}
\title{Infer file type}
\usage{
file_types()

px_file_types(fls, types = file_types())
}
\arguments{
\item{fls}{\code{character()} of file names whose types need to be
inferred based on their file extenstion.}

\item{types}{\code{data.frame} of file types. Default is
\code{file_types()}.}
}
\value{
A \code{data.frame} with the filenames and their inferred
types.
}
\description{
The \code{px_file_types()} function inferres mass spectrometry and
proteomics file types based on a currated table of file types and
associated patterns. This table can be accessed with
\code{file_types()}. See the examples below for the content of the
tables.

The file types table is generated by \code{scripts/make_file_types.R}.
}
\examples{

file_types()

px_file_types("foo")
px_file_types("foo.mzML")
px_file_types("foo.raw")
px_file_types("foo.txt")
px_file_types("foo.R")
px_file_types("foo.fasta")

px_file_types(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
}
\author{
Laurent Gatto with contributions from Dr. Samuel Wein to
identify several file types.
}
