% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{pxnodes}
\alias{pxnodes}
\title{Return the nodes of a \code{PXDataset}}
\usage{
pxnodes(pxdata, name, all = FALSE)
}
\arguments{
\item{pxdata}{An instance of class \code{PXDataset}.}

\item{name}{The name of a node.}

\item{all}{Should node from all levels be returned. Default is
\code{FALSE}.}
}
\value{
A \code{character} with XML node names.
}
\description{
Returns the node names of the underliyng XML content of an
\code{PXDataset} object, available in the \code{Data} slot. This
function is meant to be used if additional parsing of the XML
structure is needed.
}
\author{
Laurent Gatto
}

