% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{rpxCache}
\alias{rpxCache}
\title{Package cache}
\usage{
rpxCache()
}
\value{
`character(1)` with the path to the cache directory.
}
\description{
This function returns the central `rpx` cache directory.
}
\examples{

rpxCache()
}
\author{
Laurent Gatto
}
