% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{rpxCache}
\alias{rpxCache}
\alias{pxCachedProjects}
\title{Package cache}
\usage{
rpxCache()

pxCachedProjects(cache = rpxCache())
}
\arguments{
\item{cache}{Object of class \code{BiocFileCache}.}
}
\value{
\code{character(1)} with the path to the cache directory for
\code{rpxCache()} and a \code{tibble} for \code{pxCachedProjects()}.
}
\description{
Function to access the cache directory. \code{rxpCache()} returns the
central \code{rpx} cache. \code{pxCachedrojects()} prints the names of the
cached projects and invisibly returns the cache table.
}
\details{
The cache is an object of class \code{BiocFileCache}, and created with
\code{BiocFileCache::BiocFileCache()}. It can be either the
package-wide cache as defined by \code{rpxCache()} or an instaned
provided by the user.
}
\examples{

rpxCache()

pxCachedProjects()
}
\author{
Laurent Gatto
}
