% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{pxannounced}
\alias{pxannounced}
\title{Return recent PX announcements}
\usage{
pxannounced()
}
\value{
A \code{data.frame} with announcements data set
identifiers, publication dates and annoucement messages.
}
\description{
Queries the PX rss feed file for the latest PX dataset
announcements.
}
\examples{
pxannounced()
}
\author{
Laurent Gatto
}

