% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_types.R
\name{fileTypes}
\alias{fileTypes}
\alias{pxFileTypes}
\title{Infer file type}
\usage{
fileTypes()

pxFileTypes(fls, types = fileTypes())
}
\arguments{
\item{fls}{\code{character()} of file names whose types need to be
inferred based on their file extenstion.}

\item{types}{\code{data.frame} of file types. Default is
\code{fileTypes()}.}
}
\value{
A \code{data.frame} with the filenames and their inferred
types.
}
\description{
The \code{pxFileTypes()} function inferres mass spectrometry and
proteomics file types based on a currated table of file types and
associated patterns. This table can be accessed with
\code{fileTypes()}. See the examples below for the content of the
tables.

The file types table is generated by \code{scripts/make_fileTypes.R}.
}
\examples{

fileTypes()

pxFileTypes("foo")
pxFileTypes("foo.mzML")
pxFileTypes("foo.raw")
pxFileTypes("foo.txt")
pxFileTypes("foo.R")
pxFileTypes("foo.fasta")

pxFileTypes(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
}
\author{
Laurent Gatto with contributions from Dr. Samuel Wein to
identify several file types.
}
